Reworked the Smell classes a little by giving the abstract class a bit more functionality since a lot of code ended up being repeated. You can overrride any of the classes if you want more specific functionalities so it shouldn't cause an issue.
Removed few warnings by reformatting code according to Intellij's advice.

+++ src/main/java/results/SmellResult.java
@@ -67,12 +67,12 @@ public class SmellResult {
         return smellName;
     }
 
-    public int getAverageSeverity() {
-        return (int) Math.round((totalSeverity*1.0) / severityPerFile.size());
+    public double getAverageSeverity() {
+        return Math.round((totalSeverity*1.0) / severityPerFile.size());
     }
 
-    public int getAverageOccurrences() {
-        return (int) Math.round((totalOccurrences*1.0) / occurrencesPerFile.size());
+    public double getAverageOccurrences() {
+        return Math.round((totalOccurrences*1.0) / occurrencesPerFile.size());
     }
 
     public HashMap<String, Integer> getSeverityPerFile() {
